Fix conflicts or misses from merges
Lit should be a npm only feature.
Polymer should have HtmlImport version.
