+++ flow-bom/pom.xml
@@ -20,11 +20,6 @@
             </dependency>
             <dependency>
                 <groupId>com.vaadin</groupId>
-                <artifactId>flow-polymer-template</artifactId>
-                <version>${project.version}</version>
-            </dependency>
-            <dependency>
-                <groupId>com.vaadin</groupId>
                 <artifactId>flow-lit-template</artifactId>
                 <version>${project.version}</version>
             </dependency>
