+++ flow-lit-template/src/main/java/com/vaadin/flow/component/littemplate/internal/LitTemplateParserImpl.java
@@ -33,7 +33,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.vaadin.flow.component.dependency.JsModule;
-import com.vaadin.flow.component.polymertemplate.BundleParser;
 import com.vaadin.flow.function.DeploymentConfiguration;
 import com.vaadin.flow.internal.AnnotationReader;
 import com.vaadin.flow.internal.Pair;
@@ -61,7 +60,7 @@ import elemental.json.JsonObject;
  * @author Vaadin Ltd
  * @since
  *
- * @see BundleParser
+ * @see BundleLitParser
  */
 public class LitTemplateParserImpl implements LitTemplateParser {
 
@@ -132,7 +131,7 @@ public class LitTemplateParserImpl implements LitTemplateParser {
 
         Element templateElement = null;
         if (chosenDep != null) {
-            templateElement = BundleParser.parseLitTemplateElement(
+            templateElement = BundleLitParser.parseLitTemplateElement(
                     chosenDep.getFirst().getUrl(), chosenDep.getSecond());
         }
         if (templateElement != null) {
@@ -205,7 +204,7 @@ public class LitTemplateParserImpl implements LitTemplateParser {
             }
             if (!cache.containsKey(url) && jsonStats != null) {
                 cache.put(url,
-                        BundleParser.getSourceFromStatistics(url, jsonStats));
+                        BundleLitParser.getSourceFromStatistics(url, jsonStats));
             }
             return cache.get(url);
         } finally {
@@ -253,7 +252,7 @@ public class LitTemplateParserImpl implements LitTemplateParser {
     private void resetCache(String fileContents) {
         assert templateSourceslock.isHeldByCurrentThread();
         cache.clear();
-        jsonStats = BundleParser.parseJsonStatistics(fileContents);
+        jsonStats = BundleLitParser.parseJsonStatistics(fileContents);
     }
 
     private Logger getLogger() {
