+++ flow-server/src/main/java/com/vaadin/flow/component/Component.java
@@ -22,7 +22,6 @@ import java.util.Optional;
 import java.util.stream.Stream;
 import java.util.stream.Stream.Builder;
 
-import com.vaadin.flow.component.dependency.JavaScript;
 import com.vaadin.flow.component.polymertemplate.Id;
 import com.vaadin.flow.component.polymertemplate.PolymerTemplate;
 import com.vaadin.flow.dom.Element;
