+++ flow-server/src/main/java/com/vaadin/flow/component/polymertemplate/BundleParser.java
@@ -87,35 +87,6 @@ public final class BundleParser {
     private static final Pattern POLYMER_TEMPLATE_PATTERN = Pattern.compile(
             "get[\\s]*template\\(\\)[\\s]*\\{[\\s]*return[\\s]*html([\\`\\'\\\"])([\\s\\S]*)\\1;[\\s]*\\}");
 
-    /**
-     * Lit template pattern matches the template getter
-     *
-     * <pre>
-     *     render() {
-     *       return html`
-     *         &lt;style&gt;
-     *           .response { margin-top: 10px`; }
-     *         &lt;/style&gt;
-     *         &lt;paper-checkbox checked=&quot;{{liked}}&quot;&gt;I like web components.&lt;/paper-checkbox&gt;
-     *
-     *         &lt;div id=&quot;test&quot; hidden$=&quot;[[!liked]]&quot; class=&quot;response&quot;&gt;Web components like you, too.&lt;/div&gt;
-     *       `;
-     *     }
-     * </pre>
-     *
-     * <p>
-     * <code>render\(\)[\s]*\{</code> finds the template getter method
-     * <p>
-     * <code>[\s]*return[\s]*html([\`|\'|\"])</code> finds the return statement
-     * and captures the used string character
-     * <p>
-     * </p>
-     * <code>([\s\S]*)\1;[\s]*\}</code> captures all text until we encounter the
-     * end character with <code>;}</code> e.g. <code>';}</code>
-     */
-    private static final Pattern LIT_TEMPLATE_PATTERN = Pattern.compile(
-            "render\\(\\)[\\s]*\\{[\\s]*return[\\s]*html`(([^`]|\\\\.)*)`;[\\s]*\\}");
-
     private static final Pattern NO_TEMPLATE_PATTERN = Pattern.compile(
             "innerHTML[\\s]*=[\\s]*([\\`\\'\\\"])([\\s]*<dom-module\\s+[\\s\\S]*)\\1;");
 
@@ -228,44 +199,6 @@ public final class BundleParser {
                 .forEach(template::appendChild);
 
         return template;
-    }
-
-    /**
-     * Get the Lit template element for the given polymer template source.
-     *
-     * @param fileName
-     *            name of the handled file
-     * @param source
-     *            source js to get template element from
-     * @return template element or {code null} if not found
-     */
-    public static Element parseLitTemplateElement(String fileName,
-            String source) {
-        Document templateDocument = null;
-        String content = StringUtil.removeComments(source);
-        Matcher templateMatcher = LIT_TEMPLATE_PATTERN.matcher(content);
-
-        // GroupCount should be 2 as the first group contains `|'|" depending
-        // on what was in template return html' and the second is the
-        // template contents.
-        if (templateMatcher.find() && templateMatcher.groupCount() == 2) {
-            String group = templateMatcher.group(1);
-            LOGGER.trace("Found regular Lit template content was {}", group);
-
-            templateDocument = Jsoup.parse(group);
-            LOGGER.trace("The parsed template document was {}",
-                    templateDocument);
-            Element template = templateDocument
-                    .createElement(TEMPLATE_TAG_NAME);
-            Element body = templateDocument.body();
-            templateDocument.body().children().stream()
-                    .filter(node -> !node.equals(body))
-                    .forEach(template::appendChild);
-
-            return template;
-        }
-        LOGGER.warn("No lit template data found in {} sources.", fileName);
-        return null;
     }
 
     private static Element tryParsePolymer2(Document templateDocument,
