+++ flow-server/src/main/java/com/vaadin/flow/component/polymertemplate/Id.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2000-2020 Vaadin Ltd.
+ * Copyright 2000-2021 Vaadin Ltd.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
@@ -25,17 +25,17 @@ import com.vaadin.flow.component.Component;
 import com.vaadin.flow.dom.Element;
 
 /**
- * Defines the id of an element to map to inside a {@link PolymerTemplate}.
+ * Defines the id of a component or an element to map to inside a lit template.
  * <p>
  * Use this annotation with an identifier of the element which you want to refer
- * to as a value for a field inside your {@link PolymerTemplate} class.
+ * to as a value for a field inside your {@code LitTemplate} class.
  * <p>
  * Here is a Java sample:
  *
  * <pre>
  * <code>
  * &#64;Tag("details")
- * public class Details extends PolymerTemplate&lt;EditorModel&gt;{
+ * public class Details extends LitTemplate {
  *
  *      &#64;Id("name")
  *      private Div nestedDiv;
@@ -50,13 +50,13 @@ import com.vaadin.flow.dom.Element;
  *
  * <pre>
  * <code>
- * &lt;dom-module id="details"&gt;
- *   &lt;template&gt;
+ *    render(){
+ *     return html`
  *     &lt;div id='name'&gt;
  *      &lt;label&gt;Text&lt;/label&gt;
  *     &lt;/div&gt;
  *     &lt;input type="text" id='email'&gt;&lt;/div"&gt;
- *   &lt;/template&gt;
+ *     `;
  *   ....
  * </code>
  * </pre>
@@ -65,8 +65,8 @@ import com.vaadin.flow.dom.Element;
  * the element injected via <code>@Id</code> is not populated and not available
  * on the server side (it's not known). It means that <code>nestedDiv</code>
  * field value which is a <code>Div</code> component doesn't have any child on
- * the server side. Also attribute values declared on the client side are not
- * available on the server side.
+ * the server side. Attribute values declared on the client side are reflected
+ * to the server side as property values or attribute values.
  * <p>
  * You still may use {@link Component}'s or {@link Element}'s mutation methods
  * for the injected element from the server side though. E.g. you may add a
@@ -74,7 +74,7 @@ import com.vaadin.flow.dom.Element;
  * element's hierarchy in the same way as for a regular element.
  *
  * @author Vaadin Ltd
- * @since 1.0
+ * @since
  */
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.FIELD)
