+++ flow-tests/test-dev-mode/src/main/java/com/vaadin/flow/uitest/ui/template/AbstractAttachExistingElementByIdTemplate.java
@@ -18,8 +18,8 @@ package com.vaadin.flow.uitest.ui.template;
 import com.vaadin.flow.component.html.Input;
 import com.vaadin.flow.component.html.Label;
 import com.vaadin.flow.component.polymertemplate.EventHandler;
-import com.vaadin.flow.component.polymertemplate.Id;
 import com.vaadin.flow.component.polymertemplate.PolymerTemplate;
+import com.vaadin.flow.component.polymertemplate.Id;
 import com.vaadin.flow.templatemodel.TemplateModel;
 
 public abstract class AbstractAttachExistingElementByIdTemplate
