+++ flow-tests/test-npm-only-features/test-npm-general/frontend/lit/AttributeLitTemplate.js
