+++ flow-tests/test-npm-only-features/test-npm-general/frontend/lit/simple-lit-template-shadow-root.js
