+++ flow-tests/test-npm-only-features/test-npm-general/src/main/java/com/vaadin/flow/uitest/ui/littemplate/LitPerson.java
