+++ flow-tests/test-npm-only-features/test-npm-general/src/main/java/com/vaadin/flow/uitest/ui/littemplate/SimpleLitTemplateShadowRootView.java
@@ -1,3 +1,19 @@
+/*
+ * Copyright 2000-2021 Vaadin Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
+
 package com.vaadin.flow.uitest.ui.littemplate;
 
 import com.vaadin.flow.component.ClientCallable;
@@ -11,18 +27,18 @@ import com.vaadin.flow.component.polymertemplate.Id;
 import com.vaadin.flow.router.Route;
 import com.vaadin.flow.uitest.servlet.ViewTestLayout;
 
-@Tag("simple-lit-template-no-shadow-root")
-@JsModule("lit/simple-lit-template-no-shadow-root.js")
+@Tag("simple-lit-template-shadow-root")
+@JsModule("lit/simple-lit-template-shadow-root.js")
 @NpmPackage(value = "lit-element", version = "2.1.0")
-@Route(value = "com.vaadin.flow.uitest.ui.littemplate.SimpleLitTemplateNoShadowRootView", layout = ViewTestLayout.class)
-public class SimpleLitTemplateNoShadowRootView extends LitTemplate {
+@Route(value = "com.vaadin.flow.uitest.ui.littemplate.SimpleLitTemplateShadowRootView", layout = ViewTestLayout.class)
+public class SimpleLitTemplateShadowRootView extends LitTemplate {
 
     @Id
     public NativeButton mappedButton;
     @Id
     public Div label;
 
-    public SimpleLitTemplateNoShadowRootView() {
+    public SimpleLitTemplateShadowRootView() {
         getElement().setProperty("text", "Client button");
         mappedButton.setText("Server button");
         mappedButton.addClickListener(e -> {
