+++ flow-tests/test-npm-only-features/test-npm-general/src/test/java/com/vaadin/flow/uitest/ui/template/TemplateAttributeIT.java
