+++ flow-tests/test-root-context/pom.xml
@@ -29,12 +29,6 @@
             <artifactId>flow-test-common</artifactId>
             <version>${project.version}</version>
         </dependency>
-        
-        <dependency>
-            <groupId>com.vaadin</groupId>
-            <artifactId>flow-lit-template</artifactId>
-            <version>${project.version}</version>
-        </dependency>
 
         <dependency>
             <groupId>com.vaadin</groupId>
