+++ flow-tests/test-root-context/src/main/webapp/frontend/com/vaadin/flow/uitest/ui/template/template-without-shadow-root-view.html
@@ -1,0 +1,34 @@
+<!--
+  ~ Copyright 2000-2019 Vaadin Ltd.
+  ~
+  ~ Licensed under the Apache License, Version 2.0 (the "License"); you may not
+  ~ use this file except in compliance with the License. You may obtain a copy of
+  ~ the License at
+  ~
+  ~ http://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~ Unless required by applicable law or agreed to in writing, software
+  ~ distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+  ~ WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+  ~ License for the specific language governing permissions and limitations under
+  ~ the License.
+  -->
+<!-- Using an absolute URL is a bad practice, but the nesting depth here makes it incovenient traverse up with ../.. -->
+<link rel="import" href="/frontend/bower_components/polymer/polymer-element.html">
+
+<dom-module id="template-without-shadow-root-view">
+    <template>
+        <div id="content"></div>
+    </template>
+    <script>
+        class PolymerTemplateWithoutShadowRootView extends Polymer.Element {
+            static get is() { return 'template-without-shadow-root-view' }
+
+          _attachDom(dom) {
+            // Do not create a shadow root
+            this.appendChild(dom);
+          }
+        }
+        customElements.define(PolymerTemplateWithoutShadowRootView.is, PolymerTemplateWithoutShadowRootView);
+    </script>
+</dom-module>
