--- flow-tests/test-root-context/src/test/java/com/vaadin/flow/uitest/ui/littemplate/SetInitialTextLitIT.java
@@ -1,7 +1,0 @@
-package com.vaadin.flow.uitest.ui.littemplate;
-
-import com.vaadin.flow.uitest.ui.template.SetInitialTextIT;
-
-public class SetInitialTextLitIT extends SetInitialTextIT {
-
-}
